corrected task 9

+++ tasks/src/task9/Main.java
@@ -6,11 +6,10 @@ public class Main {
         Scanner num = new Scanner(System.in);
         double a, b, root;
         System.out.println("Введите первую переменную: ");
-        a = num.nextInt();
+        a = num.nextDouble();
         System.out.println("Введите вторую переменную: ");
-        b = num.nextInt();
+        b = num.nextDouble();
         root = Math.sqrt(a * b);
         System.out.println("Корень чисел " + a + " и " + b + " равен " + root);
-
     }
 }
